# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A painless HTML to PDF rendering service. Generate PDF reports and documents from HTML templates or raw HTML.",
        "title": "PdfTurtle API",
        "contact": {
            "name": "Lucas Gaitzsch",
            "email": "lucas@gaitzsch.dev"
        },
        "license": {
            "name": "AGPL-3.0",
            "url": "https://github.com/lucas-gaitzsch/pdf-turtle/blob/main/LICENSE"
        },
        "version": "1.1"
    },
    "paths": {
        "/api/pdf/from/html-bundle/render": {
            "post": {
                "description": "Returns PDF file generated from bundle (Zip-File) of HTML or HTML template of body, header, footer and assets. The index.html file in the Zip-Bundle is required",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/pdf"
                ],
                "tags": [
                    "Render HTML-Bundle"
                ],
                "summary": "Render PDF from bundle including HTML(-Template) with model and assets provided in form-data (keys: bundle, model)",
                "parameters": [
                    {
                        "type": "file",
                        "description": "Bundle Zip-File",
                        "name": "bundle",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "JSON-Model for template (only required for template)",
                        "name": "model",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "JSON-Model file for template (only required for template, will be used if model key is not provided)",
                        "name": "modelFile",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Template engine to use for template (only required for template)",
                        "name": "templateEngine",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "PDF File"
                    }
                }
            }
        },
...
    },
...
}